peripherals, storage, network, peeler, unit tests: clear the warnings
Behaviour-neutral:
- (void) on callback parameters (keyboard, rtc, via, floppy_gcr).
- deleted: voodoo2_gpu's unused v2gpu_f, floppy_swim's unused img,
  afp_desktop's never-used log category, a write-only array in the
  object_invalidate test.
- image.c: mint_random_hex_id takes char out[static 17] instead of a
  length plus an assert, so call sites are checked at compile time.
- scsi cmd_size returns size_t (6/10/12, compared with a size);
  rsrc_dcmp casts its uint16 index side.
- peeler: strncpy(dst, src, N-1) -> snprintf(dst, N, "%s", src) in
  cpt.c and sit.c; every destination was already terminated, so the
  strings are identical.
- unit tests: the rtc suite builds with -D_GNU_SOURCE like headless (it
  got strptime implicitly declared as returning int); afp and vfs test
  buffers sized so snprintf cannot truncate; the scheduler test's
  gov_drive drops an unused parameter.

+++ src/core/storage/rsrc_dcmp.c
@@ -231,7 +231,7 @@ static int vt_remember(vt_t *vt, const uint8_t *data, size_t len) {
     if (vt->size < 4)
         return -1;
     uint16_t next_idx = RD_BE16(vt->buf);
-    if (next_idx + 2 > vt->size)
+    if ((size_t)next_idx + 2 > vt->size)
         return -1; // index list would collide with the data area
     // Previous entry's offset = end of the new string.
     uint16_t prev_off = RD_BE16(vt->buf + next_idx - 2);
